% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_lb_ub}
\alias{adjust_lb_ub}
\title{Adjust lower and upper bounds of the support}
\usage{
adjust_lb_ub(lb, ub, pfd)
}
\arguments{
\item{lb}{lower bound}

\item{ub}{upper bound}

\item{pfd}{result of the partial fraction decomposition}
}
\value{
a vector of the new lower and upper bounds
}
\description{
Adjust the lower and upper bounds of the support to make sure no roots
lying within the support
}
