% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajd.sim.wh-package.R
\docType{package}
\name{ajd.sim.wh-package}
\alias{ajd.sim.wh}
\alias{ajd.sim.wh-package}
\title{ajd.sim.wh: Simulation of Affine Jump Diffusions Using Wu-Hu Method}
\description{
Simulate exactly the Heston stochastic volatility model and its affine jump diffusion extensions using the Wu-Hu method. The extensions include a) incorporating jumps in the price process, b) adding contemporaneous jumps both in the price and variance processes.
}
\author{
\strong{Maintainer}: Yan-Feng Wu \email{yanfengwu13@fudan.edu.cn} (\href{https://orcid.org/0000-0002-7105-1070}{ORCID})

}
\keyword{internal}
