% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisect.R
\name{bisect}
\alias{bisect}
\alias{bisect6}
\alias{bisect8}
\alias{bisect10}
\title{Bisection to Find the Target Quantile}
\usage{
bisect6(lb, ub, logC, U, pfd3, LB, rel_err = 1e-05)

bisect8(lb, ub, logC, U, pfd4, LB, rel_err = 1e-05)

bisect10(lb, ub, U, pfd5, LB, eps)
}
\arguments{
\item{lb}{lower bound of the quantile}

\item{ub}{upper bound of the quantile}

\item{logC}{log of the constant \eqn{C} (density normalization).}

\item{U}{target probability}

\item{pfd3, pfd4, pfd5}{solution of Partial Fraction Decomposition}

\item{LB}{lower bound of the support of the distribution}

\item{rel_err}{relative error to accept}

\item{eps}{accuracy to achieve}
}
\value{
scalar
}
\description{
Use bisection to find the target quantile
}
\examples{
lb = -0.01; ub = 0.01
}
