% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conjugate_or_equal}
\alias{conjugate_or_equal}
\alias{is.conj}
\alias{is.equal}
\title{Checking Conjugate or Equal of Two Complex Numbers}
\usage{
is.conj(z1, z2, eps = 1e-10)

is.equal(z1, z2, eps = 1e-10)
}
\arguments{
\item{z1}{complex number}

\item{z2}{complex number}

\item{eps}{relative error}
}
\value{
bool \code{TRUE} or \code{FALSE}
}
\description{
Checking Conjugate or Equal of Two Complex Numbers
}
\examples{
z1 = 1+1.000001i
z2 = 1-1.000000i
is.conj(z1, z2)
}
