% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpearson.R
\name{dpearson}
\alias{dpearson}
\alias{dpearson6}
\alias{dpearson8}
\alias{dpearson10}
\title{Un-normalized Density of Pearson Distribution}
\usage{
dpearson6(x, pfd3)

dpearson8(x, pfd)

dpearson10(x, pfd5)
}
\arguments{
\item{x}{vector of quantiles}

\item{pfd3}{list of x1(real), A1;        x2(complex), A2, B2}

\item{pfd}{Partial Fraction Decomposition encoded in a list}

\item{pfd5}{list of x1(real), A1;        x2(complex), A2, B2;
x3(complex), A3, B3}
}
\value{
scalar or vector of the density|densities
}
\description{
Un-normalized density of Pearson Distribution
}
\examples{
x = 0
}
