% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmu_to_val.R
\name{eval_mom_svcj_unc}
\alias{eval_mom_svcj_unc}
\title{Compute numerical unconditional moments for SVCJ}
\usage{
eval_mom_svcj_unc(fmu, par)
}
\arguments{
\item{fmu}{moment formula of the centralized return}

\item{par}{list of parameters}
}
\value{
scalar moment
}
\description{
Compute numerical unconditional central moments from their formulas for
the SVCJ model.
}
\examples{
# fmu.svcj.unc = fmu.svcj.unc
v0 = 0.007569; k = 3.46; theta = 0.008; sigma = 0.14
rho = -0.82; r = 0.0319; tau = 1; lmbd = 0.47; mu_b = -0.1
sigma_s = 0.0001; mu_v = 0.05; rhoJ = -0.38

mu = r - lmbd * mu_b
mu_s = log((1 + mu_b) * (1 - rhoJ*mu_v)) - sigma_s^2 / 2
h = 1

par = list(mu=mu, k=k, theta=theta, sigma=sigma, rho=rho, lmbd=lmbd,
           mu_v=mu_v, rhoJ=rhoJ, mu_s=mu_s, sigma_s=sigma_s, h=h)
moms8 = rep(0, 8) # centralized variable whose first moment = 0
for (i in 2:8) {moms8[i] = eval_mom_svcj_unc(fmu.svcj.unc[[i]], par)}
}
