% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fmu.hest}
\alias{fmu.hest}
\title{Heston SV Model Central Moment Formulas}
\format{
\subsection{\code{fmu.hest}}{

A list with 8 elements:
\describe{
\item{\code{fmu1}}{The first conditional central moment, equals 0}
\item{\code{fmu2}}{The second conditional central moment}
\item{...}{The ... conditional central moment}
\item{\code{fmu8}}{The eighth conditional central moment}
}
Each of the formula consists of a list of 9 vectors:
\describe{
\item{\code{e^{-kt}}}{Power of this term}
\item{...}{Power of this term}
\item{\code{rho}}{Power of this term}
\item{\code{num}}{Numerator for this row coefficient}
\item{\code{den}}{Denominator for this row coefficient}
}
}
}
\usage{
fmu.hest
}
\description{
The first eight conditional central moment formulas for the Heston SV model.
}
\keyword{datasets}
