% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fmu.svcj}
\alias{fmu.svcj}
\title{SVCJ Model Central Moment Formulas}
\format{
\subsection{\code{fmu.svcj}}{

A list with 8 elements:
\describe{
\item{\code{fmu1}}{The first conditional central moment, equals 0}
\item{\code{fmu2}}{The second conditional central moment}
\item{...}{The ... conditional central moment}
\item{\code{fmu8}}{The eight conditional central moment}
}
Each of the formula consists of a list of 15 vectors:
\describe{
\item{\code{e^{kt}}}{Power of this term}
\item{...}{Power of this term}
\item{\code{sigma_s}}{Power of this term}
\item{\code{num}}{Numerator for this row coefficient}
\item{\code{den}}{Denominator for this row coefficient}
}
}
}
\usage{
fmu.svcj
}
\description{
The first eight conditional central moment formulas for the SVCJ model.
}
\keyword{datasets}
