% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_p.R
\name{int_p}
\alias{int_p}
\alias{int_p8}
\alias{int_p6}
\alias{int_p10}
\title{Dynamically Compute the Cumulative Probability}
\usage{
int_p8(lb, ub, pfd4, logC, rel_err = 1e-07, max.ord = 18)

int_p6(lb, ub, pfd3, logC, rel_err = 1e-07, max.ord = 18)

int_p10(lb, ub, pfd5, eps)
}
\arguments{
\item{lb}{lower bound}

\item{ub}{upper bound}

\item{logC}{log of the constant \eqn{C} (density normalization).}

\item{rel_err}{relative error to accept}

\item{max.ord}{maximum order of the Romberg Algorithm}

\item{pfd3, pfd4, pfd5}{solution of Partial Fraction Decomposition}

\item{eps}{accuracy to achieve}
}
\value{
scalar
}
\description{
Dynamically compute the cumulative probability over an interval
}
\examples{
lb = -0.5952014; ub = 0.3030635
}
