% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_to_coef.R
\name{mom_to_coef}
\alias{mom_to_coef}
\title{Pearson Coefficients from Moments}
\usage{
mom_to_coef(moms)
}
\arguments{
\item{moms}{vector of the first six|eight|ten moments}
}
\value{
vector of the coefficients:
\itemize{
\item \eqn{\{a, c_0,\cdots, c_3\}} corresponding to the first 6 moments.
\item \eqn{\{a, c_0,\cdots, c_4\}} corresponding to the first 8 moments.
\item \eqn{\{a, c_0,\cdots, c_5\}} corresponding to the first 10 moments.
}
}
\description{
Compute the coefficients in the following by matching the first 6|8|10
moments:
\deqn{
  \frac{d p(x)} {d x} = - \frac{a+x}{c_0x^0 + \cdots + c_nx^n} p(x)
}
where \eqn{n=3, 4, 5}, \eqn{p(x)} is the (un-normalized) density of the
Pearson Distribution.
}
\examples{
moms6 = c(0.02310, 0.01915, -0.00206, 0.00199, -0.00097, 0.00077)
moms8 = c(0.0231070307, 0.0191518038, -0.0020677784, 0.0019956050,
         -0.0009780987, 0.0007711325, -0.0006488957, 0.0006454927)
moms10 = c(0.0231070307, 0.0191518038, -0.0020677784, 0.0019956050,
          -0.0009780987, 0.0007711325, -0.0006488957, 0.0006454927,
          -0.0007204021, 0.0008985861)
coef5 = mom_to_coef(moms6)
coef6 = mom_to_coef(moms8)
coef7 = mom_to_coef(moms10)
}
