% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFDecomp.R
\name{PFDecomp}
\alias{PFDecomp}
\alias{PFDecomp3}
\alias{PFDecomp4}
\alias{PFDecomp5}
\title{Partial Fraction Decomposition}
\usage{
PFDecomp3(coef5)

PFDecomp4(coef6)

PFDecomp5(coef7)
}
\arguments{
\item{coef5}{vector of the coefficients, \eqn{\{a, c_0, \cdots, c_3\}}}

\item{coef6}{vector of the coefficients, \eqn{\{a, c_0, \cdots, c_4\}}}

\item{coef7}{vector of the coefficients, \eqn{\{a, c_0, \cdots, c_5\}}}
}
\value{
solution of the partial fraction decomposition:
\itemize{
\item \code{PFDecomp3()}: list of the roots and decomposition results, four types,
i.e., \code{31}, \code{32}, \code{33}, \code{34}
\item \code{PFDecomp4()}: list of the roots and decomposition results, nine types,
i.e., \code{41}, \code{42}, ..., \code{49}
\item \code{PFDecomp5()}: list of the roots and decomposition results, twelve types,
i.e., \code{51}, \code{52}, ..., \code{512}, not fully implemented.
}
}
\description{
Partial Fraction Decomposition of the following polynomial fraction:
\deqn{
  \frac{a+x}{c_0x^0 + \cdots + c_nx^n},
}
where \eqn{n=3, 4, 5}. The decomposition depends on the roots of the
polynomial in the denominator \eqn{c_0x^0 + \cdots + c_nx^n}.
}
\examples{
coef5 = c(-0.08538000, 0.01466260, -0.07073738, 0.10918559, 0.05940828)
PFDecomp3(coef5)
}
