% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_hest.R
\name{price_hest}
\alias{price_hest}
\title{Pricing the European Call Option Under the Heston Model}
\usage{
price_hest(N, S, K, v0, tau, r, k, theta, moms, true_price)
}
\arguments{
\item{N}{number of underling asset price samples to simulate.}

\item{S}{current underling asset price.}

\item{K}{striking price.}

\item{v0}{the initial variance}

\item{tau}{maturity time, \eqn{T}.}

\item{r}{risk-less rate.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{moms}{vector of the first six|eight|ten conditional moments of the
centralized return.}

\item{true_price}{theoretical price of the option, calculable from
the Heston (1993) method.}
}
\value{
vector of pricing error and consumed time (simulation).
}
\description{
Pricing the European call option under the Heston model using the
Wu-Hu (2024) method for the simulation.
}
\details{
The Heston SV model is described by the following SDEs:
\deqn{
  d\log s(t) = (\mu - v(t)/2)dt + \sqrt{v(t)}dw^s(t),
}
\deqn{
  dv(t) = k(\theta - v(t))dt + \sigma_v\sqrt{v(t)}dw^v(t),
}
where the correlation between \eqn{w^s(t)} and \eqn{w^v(t)} is equal to
\eqn{\rho}, i.e., \eqn{cov(w^s(t), w^v(t)) = \rho t}.
The (conditional, with \eqn{v_0} given) return over period \eqn{[0,t]}
is defined as \eqn{y_t \equiv \log s(t) - \log s(0)}
which is decomposed as
\deqn{
  y_t = (\mu - \theta/2)t - \beta_t (v_0 -\theta) + \overline{y}_t|v_0,
}
where \eqn{\beta_t = (1-e^{-kt})/(2k)} and the centralized conditional
return
\deqn{
  \overline{y}_t|v_0 = \frac{\sigma_v}{2k}e^{-kt}I\!E_t +
  \left(\rho - \frac{\sigma_v}{2k}\right)I_t +
  \sqrt{1-\rho^2}I_t^{*}.
}
Given the initial variance \eqn{v_0}, the function generate samples of
the centralized conditional return \eqn{\overline{y}_t|v_0}.
}
\examples{
# setting 1
S = 100; K = 100; v0 = 0.010201; k = 6.21; theta = 0.019; sigma = 0.61
rho = -0.7; r = 0.0319; tau = 1; true_price = 6.8061

par_hest = list(v0=v0, k=k, theta=theta, sigma=sigma, rho=rho, h=tau)
moms = rep(0, 8) # centralized variable, whose first moment always = 0
for (i in 2:8) {moms[i] = eval_mom_hest(fmu.hest[[i]], par_hest)}

price_hest(10000, S, K, v0, tau, r, k, theta, moms, true_price)

# setting 2
S = 100; K = 100; v0 = 0.09; k = 2.00; theta = 0.09; sigma = 1.00
rho = -0.3; r = 0.05; tau = 5
true_price = 34.9998

par_hest = list(v0=v0, k=k, theta=theta, sigma=sigma, rho=rho, h=tau)
moms = rep(0, 8) # centralized variable, whose first moment always = 0
for (i in 2:8) {moms[i] = eval_mom_hest(fmu.hest[[i]], par_hest)}

price_hest(10000, S, K, v0, tau, r, k, theta, moms, true_price)

}
\seealso{
\code{\link[=rpearson]{rpearson()}}
}
