% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_svj.R
\name{price_svj}
\alias{price_svj}
\title{Pricing the European Call Option Under the SVJ Model}
\usage{
price_svj(N, S, K, v0, tau, r, k, theta, lmbd, mu_b, sigma_s, moms, true_price)
}
\arguments{
\item{N}{number of underling asset price samples to simulate.}

\item{S}{current underling asset price.}

\item{K}{striking price.}

\item{v0}{initial variance, \eqn{v_0}.}

\item{tau}{maturity time, \eqn{T}.}

\item{r}{risk-less rate.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{lmbd}{parameter \eqn{\lambda}, arrival rate of jumps in the
underling asset price process.}

\item{mu_b}{parameter \eqn{\bar{\mu}}, related to the mean of the jumps,
\eqn{\mu_s = \log (1+\bar{\mu}) - \sigma_s^2/2}, see Broadie-Kaya (2006).
Meanwhile the expected return per unit time of the SVJ model becomes
\eqn{r - \lambda\bar{\mu}}.}

\item{sigma_s}{parameter \eqn{\sigma_s}, standard deviation of the jumps,
see Broadie-Kaya (2006).}

\item{moms}{vector of the first eight conditional moments of
the centralized return of the Heston SV model}

\item{true_price}{theoretical true price of the option, calculable by
the Bates (1996) method.}
}
\value{
vector of pricing error and consumed time (simulation).
}
\description{
Pricing the European call option under the SVJ model using the
Wu-Hu (2024) method for the simulation.
}
\examples{
S = 100; K = 100; v0 = 0.008836; k = 3.99; theta = 0.014; sigma = 0.27
rho = -0.79; r = 0.0319; tau = 5; lmbd = 0.11; mu_b = -0.12
sigma_s = 0.15; true_price = 20.1642

par_hest = list(v0=v0, k=k, theta=theta, sigma=sigma, rho=rho, h=tau)
moms = rep(0, 8) # centralized variable, whose first moment always = 0
for (i in 2:8) {moms[i] = eval_mom_hest(fmu.hest[[i]], par_hest)}

N = 10000
price_svj(N, S, K, v0, tau, r, k, theta, lmbd, mu_b, sigma_s,moms,true_price)
}
\seealso{
\code{\link[=r_svj]{r_svj()}}
}
