% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printf}
\alias{printf}
\title{printf function}
\usage{
printf(fmt, ...)
}
\arguments{
\item{fmt}{format like that in \link{sprintf}}

\item{...}{additional parameters}
}
\value{
None
}
\description{
printf similar to printf in C languange
}
\examples{
fmt = "pi = \%f"
printf(fmt, pi)
}
