% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_srjd.R
\name{r_srjd}
\alias{r_srjd}
\title{Simulation of the Square-Root Jump diffusion}
\usage{
r_srjd(v0, k, theta, sigma, lmbd, mu_v, h)
}
\arguments{
\item{v0}{the initial state.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{volatility of the SRJD.}

\item{lmbd}{jump intensity \eqn{\lambda}.}

\item{mu_v}{mean of the exponential jump distribution.}

\item{h}{terminal time.}
}
\value{
scalar
}
\description{
Generate samples from the Square-Root Jump Diffusion (SRJD).
}
\examples{
v0 = 0.007569
k = 3.46; theta = 0.008; sigma = 0.14; lmbd = 0.47; mu_v = 0.05; h = 1

r_srjd(v0, k, theta, sigma, lmbd, mu_v, h)
}
