% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_svcj.R
\name{r_svcj}
\alias{r_svcj}
\title{Simulate Partially Centralized Conditional Returns of SVCJ}
\usage{
r_svcj(n, moms)
}
\arguments{
\item{n}{number of return samples to simulate}

\item{moms}{vector of the first six|eight|ten conditional moments of the
centralized return.}
}
\value{
vector of the partially centralized conditional return samples
}
\description{
Simulate the partially centralized conditional return samples of the SVCJ
model by the superposition of
\itemize{
\item approximated Pearson Distribution which matches the first six|eight|ten
conditional (partially) central moments of the unknown true SVVJ return
distribution, see the Details. And
\item exact contemporaneous jump in the price (not centralized)
}
}
\examples{
S = 100; K = 100; v0 = 0.007569; k = 3.46; theta = 0.008; sigma = 0.14
rho = -0.82; r = 0.0319; tau = 1; lmbd = 0.47; mu_b = -0.1
sigma_s = 0.0001; mu_v = 0.05; rhoJ = -0.38; true_price = 6.8619

n = 100
mu = r - lmbd * mu_b
mu_s = log((1 + mu_b) * (1 - rhoJ*mu_v)) - sigma_s^2 / 2
h = 1

par = list(v0=v0, mu=mu, k=k, theta=theta, sigma=sigma, rho=rho, lmbd=lmbd,
           mu_v=mu_v, rhoJ=rhoJ, mu_s=mu_s, sigma_s=sigma_s, h=h)
moms = rep(0, 8) # centralized variable whose first moment = 0
for (i in 2:8) {moms[i] = eval_mom_svcj(fmu.svcj[[i]], par)}

Y = r_svcj(n, moms)
}
