% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_svj.R
\name{r_svj}
\alias{r_svj}
\title{Simulate Partially Centralized Conditional Returns of SVJ}
\usage{
r_svj(n, tau, lmbd, mu_b, sigma_s, moms)
}
\arguments{
\item{n}{number of return samples to simulate}

\item{tau}{time difference or maturity term}

\item{lmbd}{arrival rate \eqn{\lambda} of jumps in the price process}

\item{mu_b}{parameter \eqn{\bar{\mu}}, related to jump mean through
\eqn{\mu_s = \log(1+\bar{\mu}) - \sigma_s^2/2}}

\item{sigma_s}{standard deviation of the normally distributed jumps}

\item{moms}{vector of the first six|eight|ten conditional moments of
the centralized return of the Heston SV model}
}
\value{
vector of \eqn{\overline{y}_{hest,t}|v_0 + I\!Z_t} samples.
}
\description{
Simulate partially centralized conditional return samples of the SVJ model by
the superposition of
\itemize{
\item approximated Pearson Distribution which matches the first six|eight|ten
conditional moments of the centralized return of the Heston SV model, and
\item exact jump distribution (not centralized).
}
}
\details{
The SVJ model is described by the following SDEs:
\deqn{
  d\log s(t) = (\mu - v(t)/2) dt + \sqrt{v(t)}dw^s(t) + dz(t),
}
\deqn{
  dv(t) = k(\theta - v(t))dt + \sigma_v\sqrt{v(t)}dw^v(t)
}
where \eqn{z(t)} is a Compound Poisson process with constant arrival
rate \eqn{\lambda}. The (conditional, with \eqn{v_0} given) return over
period \eqn{[0,t]} is defined as \eqn{y_t \equiv \log s(t) -\log s(0)}
which is decomposed as
\deqn{
  y_t = y_{hest,t} + I\!Z_t,
}
where \eqn{y_{hest,t}} denotes the return of the Heston model and
\eqn{I\!Z_t\equiv \int_0^tdz(u)}. And return of the Heston model is
decomposed as
\deqn{
  y_{hest,t} = (\mu - \theta/2)t - \beta_t (v_0 -\theta) +
  \overline{y}_{hest,t}|v_0,
}
See \code{\link[=rpearson]{rpearson()}} for
generating \eqn{\overline{y}_{hest,t}|v_0} samples from the Heston
SV model.
The function \code{\link[=r_svj]{r_svj()}} generates the
\eqn{\overline{y}_{hest,t}|v_0 + I\!Z_t} samples.
}
\examples{
n = 100
# moms6 = c(0.0231, 0.0191, -0.0021, 0.0020, -0.0010, 0.0008)
v0 = 0.008836; k = 3.99; theta = 0.014; sigma = 0.27; rho = -0.79
lmbd = 0.11; mu_b = -0.12; sigma_s = 0.15; tau = 5

par = list(v0=v0, k=k, theta=theta, sigma=sigma, rho=rho, h=tau)
moms8 = rep(0, 8) # centralized variable whose first moment = 0
for (i in 2:8) {moms8[i] = eval_mom_hest(fmu.hest[[i]], par)}

Y = r_svj(n, tau, lmbd, mu_b, sigma_s, moms8)
}
\seealso{
\code{\link[=rpearson]{rpearson()}}
}
