% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpearson.R
\name{rpearson}
\alias{rpearson}
\title{Simulate Samples from Pearson Distribution}
\usage{
rpearson(n, moms)
}
\arguments{
\item{n}{number of return samples to simulate}

\item{moms}{vector of the first eight conditional central moments of
the return}
}
\value{
vector of centralized conditional return samples,
\eqn{\overline{y}_t|v_0}.
}
\description{
Simulate samples from the approximated Pearson Distribution which matches
the first eight moments of the unknown true return distribution.
}
\examples{
v0 = 0.010201; k = 6.21; theta = 0.019; sigma = 0.61; rho = -0.7
r = 0.0319; tau = 1

par = list(v0=v0, k=k, theta=theta, sigma=sigma, rho=rho, h=tau)

moms8 = rep(0, 8) # centralized variable whose first moment = 0
for (i in 2:8) {moms8[i] = eval_mom_hest(fmu.hest[[i]], par)}

n = 100
Y = rpearson(n, moms8)
}
