% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_refine.R
\name{bound_refine}
\alias{bound_refine}
\alias{bound}
\alias{refine}
\title{Bound and Refine to Find the Target Quantile}
\usage{
bound_refine(UC, px, dx, x, pfd, pfd_tp)

bound(UC, px)

refine(UC, lb, ub, px_lb, dx_lb, pfd, pfd_tp)
}
\arguments{
\item{UC}{random number U times the normal constant C}

\item{px}{vector of un-normalized cumulative probability}

\item{dx}{vector of un-normalized density}

\item{x}{vector of quantiles}

\item{pfd}{list of Partial Fraction Decomposition results}

\item{pfd_tp}{type of \code{pfd}:
\itemize{
\item 3: result of \code{\link[=PFDecomp3]{PFDecomp3()}}
\item 4: result of \code{\link[=PFDecomp4]{PFDecomp4()}}
\item 5: result of \code{\link[=PFDecomp5]{PFDecomp5()}}
}}

\item{lb}{lower bound of the target quantile}

\item{ub}{upper bound of the target quantile}

\item{px_lb}{un-normalized cumulative probability of the lower bound \code{lb}}

\item{dx_lb}{un-normalized density of the lower bound \code{lb}}
}
\value{
\itemize{
\item \code{bound_refine}: scalar, quantile point \eqn{x} such that \eqn{P(X\le x)=U}.
\item \code{bound}: vector (lower_bound_index, upper_bound_index, target_index).
\item \code{refine}: scalar as that in \code{bound_refine}.
}
}
\description{
Bound and refine the target quantile, \code{bound_refine}.
Get the lower and upper bound of the target quantile, \code{bound}.
Refine within the bounds to find the target quantile, \code{refine}.
}
\examples{
U = runif(1)
}
